pyproject.toml
[tool.poetry]
name = "fggstore-card-bot"
version = "2.0.0"
description = "Professional Telegram bot for generating PlayStation Store card images"
authors = ["FGGSTORE"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^20.6"
Pillow = "^10.0.0"
python-dotenv = "^1.0.0"
arabic-reshaper = "^3.0.0"
python-bidi = "^0.4.2"

[tool.poetry.group.dev.dependencies]
black = "^23.0.0"
mypy = "^1.0.0"
pytest = "^7.0.0"
pytest-cov = "^4.0.0"
pytest-asyncio = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
